+++ app/src/main/java/com/iosharp/android/ssplayer/tasks/FetchLoginInfoTask.java
@@ -18,11 +18,7 @@ import com.squareup.okhttp.Response;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
 import java.net.URL;
 
 public class FetchLoginInfoTask extends AsyncTask<Void, Void, Void> {